# Switch compression package (3x faster)