--- a/go.mod
+++ b/go.mod
@@ -4,6 +4,7 @@ go 1.13
 
 require (
 	github.com/gin-gonic/gin v1.5.0
+	github.com/klauspost/compress v1.11.3
 	github.com/signalsciences/ac v1.1.0
 	github.com/stretchr/testify v1.4.0
 )