--- a/handler_test.go
+++ b/handler_test.go
@@ -113,7 +113,7 @@ func TestNewHandler_Checks(t *testing.T) {
 
 	assert.Panics(t, func() {
 		NewHandler(Config{
-			CompressionLevel: -3,
+			CompressionLevel: -4,
 			MinContentLength: 100,
 		})
 	})
@@ -281,6 +281,39 @@ func TestGinWithDefaultHandler(t *testing.T) {
 			body, err := ioutil.ReadAll(reader)
 			require.NoError(t, err)
 			require.True(t, bytes.HasPrefix(body, []byte(seq)))
+		})
+	}
+}
+
+func TestGinWithLevelsHandler(t *testing.T) {
+	for i := Stateless; i < 10; i++ {
+		var seq = "level_" + strconv.Itoa(i)
+		i := i
+		t.Run(seq, func(t *testing.T) {
+			g := newEchoGinInstance(bigPayload, NewHandler(Config{
+				CompressionLevel: i,
+				MinContentLength: 1,
+			}).Gin)
+
+			var (
+				w = httptest.NewRecorder()
+				r = httptest.NewRequest(http.MethodPost, "/", strings.NewReader(seq))
+			)
+
+			r.Header.Set("Accept-Encoding", "gzip")
+			g.ServeHTTP(w, r)
+
+			result := w.Result()
+			require.EqualValues(t, http.StatusOK, result.StatusCode)
+			require.Equal(t, "gzip", result.Header.Get("Content-Encoding"))
+			comp, err := ioutil.ReadAll(result.Body)
+			require.NoError(t, err)
+			reader, err := gzip.NewReader(bytes.NewReader(comp))
+			require.NoError(t, err)
+			body, err := ioutil.ReadAll(reader)
+			require.NoError(t, err)
+			require.True(t, bytes.HasPrefix(body, []byte(seq)))
+			t.Logf("%s: compressed %d => %d", seq, len(body), len(comp))
 		})
 	}
 }