Use `github.com/klauspost/compress` package for compression.

Benchmarks did not actually compress since the header was carried over between calls.

Before/after are both with the benchmark fix in place.

```
λ benchcmp before.txt after.txt
benchmark                                          old ns/op     new ns/op     delta
BenchmarkSoleGin_SmallPayload-12                   279           284           +1.79%
BenchmarkGinWithDefaultHandler_SmallPayload-12     653           662           +1.38%
BenchmarkSoleGin_BigPayload-12                     271           284           +4.80%
BenchmarkGinWithDefaultHandler_BigPayload-12       83649         25616         -69.38%
```

I believe the last benchmark is the only one actually to be compressing, correct?

I didn't go out of my way to clean up the benchmark since the difference is rather obvious.

The compression ratio is likely a bit less, but among many other changes the default settings have been tweaked to give the most reasonable speed/compression tradeoff.