@nanmu42 2020-11-29
> Benchmarks did not actually compress since the header was carried over between calls.

My bad, thanks for fix that.

> I believe the last benchmark is the only one actually to be compressing, correct?

Yes, it is.

> I didn't go out of my way to clean up the benchmark since the difference is rather obvious.

I will clean it up. :rofl: 

> The compression ratio is likely a bit less, but among many other changes the default settings have been tweaked to give the most reasonable speed/compression tradeoff.

Generally speaking, is `CompressionLevel = 6` a good default setting?